praefect: Drop sha256 tag
The tests in the subcmd_configuration_validate_test.go file don't
setup repositories or interact with any existent, so it doesn't
need to keep the tag and "migrate" it. The only thing that is
required is to drop the tag from the main_test.go file as well
because of the testhelper.Run(m) call used there and required
when testhelper.TempDir(t) is called.