cli/gitaly: Update storage flag default value
Currently the `storage` flag used by the `hooks set` command is set to
"default". A better default would be to extract the storage name from
the config. This change updates the `storage` flag default to use the
config storage if there is only a single storage specified. If multiple
storages are specified in the config and no flag provided an error is
returned.