backup: Extract GetWriter from sink
The Write methods in each implementation are taking a Reader which then
has io.Copy called on to move the data to the sink. This results both in
duplication in the implementation, but also sometimes forces callers to
use io.Pipe to convert between a io.Writer and io.Reader. Instead here
we just return the native type and rely on the caller to clean up
properly.