git/housekeeping: Improve OptimizeRepository tests for pooled repos
The tests for OptimizeRepository with pooled repositories are cluttered
around two different files. The reason for this was so that we can make
use of the `objectpool` package to create the object pool. But this is
not really needed as we can easily create the gitalternates link by
ourselves.

Merge the tests into our normal table-driven tests and add additional
testcases.