git/housekeeping: Improve error logging when git-repack(1) fails
It's hard to deduce why exactly git-repack(1) fails as we don't
propagate its standard error via the error metadata. Refactor the code
to use the structerr package so that we can propagate it.

While at it, stop embedding `exec.ExitError`s directly. These errors are
almost impossible to test for and ultimately provide no useful info
except for their error code. This is done to improve testability in an
upcoming commit.