git/housekeeping: Reintroduce geometric repacking strategy
This commit reintroduces the geometric repacking that has originally
been introduced via 12bde1a93 (git/housekeeping: Introduce geometric
repacking strategy, 2023-03-30). This commit had to be reverted via
b76245f82 (Revert "git/housekeeping: Introduce geometric repacking
strategy", 2023-03-31) because it caused downstream failures in the
Rails test suite.

The root cause of this error is that geometric repacking does not
correctly handle repositories with an alternate because git-repack(1)
did not honor whether packfiles are local or not. This can lead to
git-repack(1) dying when either the repository has all of its objects
deduplicated via an object pool, or when the repository and its object
pool have the exact same packfile.

We are going to work around this failure in the next commit, so with
this revert we can demonstrate that we do indeed detect the failure mode
via our tests now.

Changelog: added