localrepo: Remove function to calculate repository size
The `localrepo` structure exposes a function to calculate the on-disk
size of a repository. We have since decided to take an alternative
approach via the `RepositoryInfo` RPC that utilizes the `git/stats`
package in order to obtain both the total repository size as well as
sizes of some of the underlying data structures.

Let's remove the unused function. We ain't gonna need it.

Changelog: removed