Add RequireObjects for asserting object database content
This commit adds RequireObjects test helper for asserting the object
database content. We already have helpers for checking whether a
particular object exists or not. Using those helpers however is only
useful for checking objects that one expects to exist or not exist
in the repository which easily allows for unexpected objects to
be missed in the assertions. RequireObjects asserts the entire state
which means it won't miss anything.