Expose logic to quarantine a repository
Quarantine package currently only supports quarantining a repository
with a new temporary directory. The TransactionManager will need to
apply a quarantine directory on a localrepo it already has. This commit
exposes the Apply function from the quarantine package to apply a given
directory as a quarantine on a given gitalypb.Repository. This will
later be wired into localrepo so it can return a quarantined version
of itself for use.

The localrepo using tests were moved to an external test package as
otherwise they trigger a cyclic import when localrepo imports this
logic.