Support quarantining a localrepo.Repo
Currently there's no way to quarantine a localrepo instance one
already has available. Applying a quarantine to an existing Repo
instance is preferable to building a new one with a quarantined
GitRepo as the latter would require the call site to know how to
build a localrepo instance. This commit implements Quarantine
method on the localrepo to apply a quarantine directory to it. It
return a new localrepo instance which refers to the same repository
although with the quarantine directory applied.