Move FixDirectoryPermissions to perm package
In order to support quarantining a localrepo.Repo instance, we'd
have to export the quarantining logic from the 'quarantine' package
to avoid reimplementing it. If we do that, and import 'quarantine'
package in the 'localrepo' package, we'll end up with a cyclic import
as the 'housekeeping' package uses 'localrepo', and 'tempdir' package
uses 'housekeeping'. Let's break this import cycle by moving the shared
functionality from the 'housekeeping' package to 'perm' package that can
be imported by 'tempdir' without ending up in a cycle.