ssh: Fix fragile timeout in SSHUploadPack tests to use manual tickers
In order to test whether we correctly honor the configured timeout for
the initial negotiation phase in SSHUploadPack we configure a very low
timer. This is fragile as it ultimately still relies on time-sensitive
logic.

Refactor the test to instead use a manual ticker so that we can properly
synchronize the logic and inject a timeout at exactly the point in time
we intend to.