Create TransactionManager's directories in initialize
TransactionManager stores state in directories in the repositories.
Currently those directories are not guaranteed to exist when the
TransactionManager is running. This forces the code to handle the
fact. This commit introduces createDirectories which creates the
expected directories if they don't exist when initializing. We then
simplify locations that had to handle the case when an expected
directory did not exist.

For now there is just the hooks directory but we'll soon have a
directory for pack files as well. The tests are updated to assert
the default state automatically so we don't have to repeat it
unnecessarily.