Generalize RequireHooks to RequireDirectoryState
The TransactionManager is storing state in a directory inside the
repository. For now, it's just storing the hooks but soon it will
also store the pack files. This commit generalizes the assertion of
the hooks to assert the directory's state as a whole. This will
later enable us to assert the state of the pack files as well.