Merge TransactionManager test setup into a struct
TransctionManager's tests have a function to setup the test state.
It's currently returning each individual data point needed for the
tests as a separate return value. This commit merges these values
into a struct which is easier to extend. We'll use this soon to
return pack files along the test commits.