Support parsing content in RequireDirectoryState
RequireDirectoryState allows for asserting that a directory matches
the expected state. Right now it only supports matching file content
against expected bytes. This is inconvenient as it can be difficult
to make sense of binary file formats on failure and say what exactly
is wrong without looking at the file itself. Further, some files may
be physically different but semantically the same. An example of such
a file would be Git's pack files, where the physical format may change
for example depending on the delta compression but the actual contained
objects are the same.

This commit adds support for parsing the file's content with a parser
passed in the assertion. The parser takes the file's contents and returns
a parsed object of any type. The parsed type is ultimately asserted for
equality with the expected content. With this in place, we can for example
assert the content of pack files rather than than the physical file bytes.

Missing tests are added for the function.