commit: Pass repository executor into `newLSTreeParser()`
We're about to introduce support for the SHA256 object format for
ListLastCommitsForTree, which requires access to a repository executor
in order to derive the object format that is in use. And while the
caller already has the localrepo around, we do not currently have it
accessible in `newLSTreeParser()`.

Refactor `newLSTreeParser()` to accept a repository executor so that we
can avoid having to recreate it.