commit: Implement SHA256 compatibility in ListLastCommitsForTree
The ListLastCommitsForTree RPC sets up a git-ls-tree(1) parser that
needs to know about the specific object hash that is being used in the
repository in order to correctly parse the tree. Right now we have the
SHA1 object hash hardcoded here.

Fix the code to detect the object format dynamically. This makes the RPC
compatible with SHA256-based repositories.

Changelog: added