commit: Detect early whether the repo exists in ListLastCommitsForTree
We don't explicitly verify whether the repository handed to us in the
ListLastCommitsForTree RPC actually exists or not. This means that
adding new code paths that access that repository is very likely to also
change the client-observable generated error.

Add an explicit check up front so that we can guarantee that the
repository exists throughout the RPC.