git/housekeeping: Improve metrics when repacking objects fails
When repacking objects fails during housekeeping we report a bunch of
metrics relating to this as failed. We completely ignore the repack
configuration though that led to the failure, which means that we would
e.g. report both a "full" and an "incremental" repack as failed even
though we actually performed a "cruft" repack. Furthermore, we do not
take into account whether we actually tried to write bitmaps or the
multi-pack index but also always report them as failed. This means that
our metrics can be misleading at times, and that we cannot see whether a
specific type of packing objects is more likely to fail than others.

Fix this error by considering the repack configuration when reporting
error metrics.

Changelog: fixed