localrepo: Turn TreeEntry into a tree data structure
Currently, a TreeEntry is a flat data structure in the sense that it
represents a single entry but does not have any children. A git tree
object has children though.

Change this by adding an Entries member to the TreeEntry struct. This
way, we can represent a real tree data structure.

Also, add a GetTree() function that can get an entire tree. We can
consolidate ListEntries() into GetTree().