Dangerfile: Add group::gitaly::git and group::gitaly::cluster
These two new groups are going to replace gitaly::git and
gitaly::cluster. See https://gitlab.com/gitlab-com/www-gitlab-com/-/merge_requests/122793
for details. Add them to the INHERITABLE_LABELS so they get
inherited from issues.

Also the group::gitaly label can also be inherited, in the situation
that the issue is a group::gitaly issue but a community contributor or
someone on another team ends up contributing to it.