gitaly: Generalize repository ID generation
Currently `getRepositoryID()` requires a `repository` as an argument. In
a future commit a repository ID will also need to be generated directly
from a `repository.GitRepo`. Since `repository` already implements
`repository.GitRepo` and `getRepositoryID` does not rely on anything
additional, the argument type is change from `repository` to
`repository.GitRepo`.