gitaly: Introduce transaction finalizers
Currently `TransactionManager` is explicitly started before generating
transactions. It also must be explicitly stopped once all transactions
are complete. A future commit will introduce `PartitionManager` which
will manage the lifecycle for all instances of `TransactionManager`. In
order to know that a transaction is finished a signal must be sent back
to `PartitionManager`.

This change introduces `transactionFinalizer` to `TransactionManager`
which executes on transaction commit and rollback. This can be leveraged
by the `PartitionManager` to perform `TransactionManager` cleanup.