Generalize RequireObjects to ListObjects
RequireObjects currently asserts that the repository's object database
contains the expected objects. This commit extracts the object listing
code from the function to ListObjects. This is more flexible as one
can then do whatever is needed with the given objects. The listing code
is changed to include the empty tree oid. It doesn't make sense to filter
it out in the more general context.

As RequireObjects is mostly now a wrapper around ListObjects, it is
unnecessary as one can just pass the objects with the expected objects
to require.ElementsMatch as usual.