git/housekeeping: Extract constants for full-repack cooldown periods
We're using cooldown periods to figure out whether it's time for another
full repack or not. So we will only consider doing a full repack in case
the previous full repack is longer ago than the cooldown period.

This is a value we can and want to fine-tune as we gain experience with
geometric repacking. So let's these into a constant that we can use in
our tests so that we don't have to adjust them every time we change the
cooldown period.