ci: Disable some jobs for tagged builds
At the moment, when a tag is pushed, the tag is created with 'ci skip'.
This change is made in release-tools@6045cf49 (Don't skip CI builds for
Omnibus and CNG, 2020-08-21) because we don't want tests to run on
tagged builds.

With this change we disable some jobs for tagged builds in this CI
configuration. This makes it possible to remove 'ci skip' from tags and
have tagged builds publish the proto gem.