ref: Refactor ListTagNamesContainingCommits tests
Refactor tests for the ListTagNamesContainingCommits RPC to conform to
our modern best practices. Most importantly, refactor it so that we
don't use a seed repository anymore but instead generate test data at
runtime. This also allows us to be more thorough with our testing where
we don't need to use a subset-test anymore, but can instead test for
equivalence of expected and actual results.