ref: Support SHA256 in RPCs that search for refs containing commits
The RPCs that search for either branches or tags that contain a specific
commit are using hardcoded SHA1 object hash to parse arguments passed by
the caller. This of course breaks with repositories that use SHA256 as
the object format.

Detect the object hash that needs to be used at runtime to fix support
for SHA256. Enable testing the RPCs with SHA256.

Changelog: added