Write-ahead log repository deletions
Repository deletions need to be write-ahead logged as well. This
ensures their atomicity and that they can be replicated later as
part of the log.

Once a repository deletion has been commited, all subsequent Begin
and Commit calls will fail with a 'repository not found' error. The
repository is logically deleted but not yet physically. The physical
deletion needs to wait for open transactions to finish so we don't
remove the files they are operating on.

For now, it's possible to set other updates in the Transaction even
if it ultimately removes the repository. Technically this is fine but
it's a bit non-sensical and we don't have a use case for it in Gitaly.
We'll probably later improve the interface by splitting out different
transaction types so UpdateReferences() can't be called on the same
transaction that deletes the repository.