Pass storage and relative path directly to TransactionManager
TransactionManager is currently retrieving the repository path through
the localrepo.Repo. This won't work in the future anymore once we WAL
repository creations and deletions. Repo.Path() checks whether the
directory is a git repository, which won't be the case if the repository
was deleted or hasn't been created yet. We need the path when initializing
the manager so we can load its state correctly. We're taking the storage path
and relative path instead of a GitRepo as it works better in the context
of the TransactionManager:

1. The TransactionManager always works within a single storage, given it
   operates on a single repository. The typical interface of using a Locator
   to find the path of a repository is thus needlessly complex, and would force
   error handling to check a non-existent storage is not accessed. Taking the path
   directly doesn't require unnecessary error handling here, and forces the caller
   to handle it.
2. For most cases, just taking the repository's absolute path would be enough
   However, we also need an identifier to use in the database keys. Currently
   the identifier is <storage>:<relative_path>. Each storage will be an independent
   failure domain and will have their own database instance. Given that, a
   relative path is enough to uniquely identify a repository within a storage.
   We also don't have a better identifier available yet so we use a relative path.
   After the RenameRepository is removed, the relative path will be stable and okay
   to use as an identifier.

The PartitionManager is extended to take in the configured storages so it can pass
the required parameters when constructing the TransactionManagers.