Prevent transaction beginning if initialization fails
Begin currently waits for the TransactionManager to initialize but
it doesn't check whether the initialization was successful. As the
transaction will not anyway succeed if the initialization failed,
return an appropriate error from Begin instead. This way the
transaction doesn't end up doing any useless work and the error
message will be clearer than having a random failure at some point
during the transaction.