git/stats: Encapsulate functionality for last-full-repack timestamps
The knowledge for how to read and write the last-full-repack timestamps
is scattered across the git/housekeeping and git/stats packages. While
this is fine right now because there's only a single reader and a single
writer, we're about to add a second writer.

Refactor the code to provide encapsulated functions that read and update
the timestamp. Add tests to verify that things work as expected.