git/housekeeping: Introduce repack strategy to pack all loose objects
While all of our strategies for packing objects will pack loose objects
that exist in the repository, none of them will _only_ pack loose
objects:

    - The full repacking strategies will pack all objects into a single
      packfile.

    - The geometric repacking strategy will include all loose objects,
      but potentially also merges packfiles to restore the geometric
      sequence.

    - The incremental repacking strategy will not pack unreachable loose
      objects and thus needs to compute reachability of objects.

In the case where we know that a repository is well-packed but has a
bunch of loose objects that we want to pack all of these strategies are
needlessly expensive.

Introduce a new strategy that allows us to pack all not-yet-packed loose
objects into a single packfile regardless of their reachability. Because
git-repack(1) does not offer such a mode for us, we instead need to
implement it via calls to git-pack-objects(1) and git-prune-packed(1).

This new strategy will be used in a follow-up commit.