repoutil: Copy GetCustomHooks from repository service
Soon we will be implementing server-side backups which need to be able
to archive custom hooks without calling an RPC. This copies the
GetCustomHooks RPC implementation so that it can then be used for
backups. Once this is complete the original implementation will be
removed.

This functionality is not appropriate for localrepo because it is built
on top of git and isn't part of git itself. It also doesn't seem like a
good fit for "internal/gitaly/hook" because that package is about
running hooks rather than managing custom hook scripts.