Makefile: Clarify variables that specify test report paths
The variables `TEST_REPORT` and `TEST_FULL_OUTPUT` allow the developer
to specify where the test report and the full test output shall be
written to. It's not immediately obvious what this is though:

    - `TEST_REPORT` specifies the location of the JUnit-style test
      report.

    - `TEST_FULL_OUTPUT` specifies the location of the JSON-style test
      report as it would be written by `go test -json`.

Rename the variables to `TEST_JUNIT_REPORT` and `TEST_JSON_REPORT` to
hopefully clarify what their respective intent is. Adjust their paths to
match.