git/housekeeping: Reenable geometric repacking with alternates
When we first tried to roll out geometric repacking we quickly hit
issues with them not working in the context of repositories which are
connected to an alternate object database. The following investigation
surfaced different errors in git-repack(1), git-pack-objects(1) and
git-multi-pack-index(1) when alternates come into play. We have thus
adapted the housekeeping strategy to only perform geometric repacking
when the repository is not connected to an alternate object database.

Since then we have upstreamed a set of fixes which have been merged via
de56e80363 (Merge branch 'ps/fix-geom-repack-with-alternates' into next,
2023-04-18). With these fixes applied, there are currently no bugs that
would keep us from enabling geometric repacks in pooled repositories.

Enable geometric repacking in repositories which are part of an object
pool when we detect that the current Git version has the fixes.

Changelog: changed