git/housekeeping: Perform validation of config early in RepackObjects
The configuration passed to `RepackObjects()` can be inconsistent, e.g.
when trying to write a bitmap for an incremental repack that is not also
updating multi-pack-indices. We perform this validation after we have
already performed some of the business logic to figure out which flags
to pass to git-repack(1). This is confusing though and hurts readability
of this function.

Reorder the checks so that we perform them up-fornt before we execute
any business logic.